cleanup(netxlite): add file for quirky behavior required by netx
There are a bunch of quirks in the codebase that are there because the
./legacy/netx package assumes them. So, in light of the fact that I'm
going to modify how we construct HTTP clients, I feel a bit safer if we
move what I should not touch in a different place.

Part of https://github.com/ooni/probe/issues/2531
